services/notification-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-socketio==5.11.0
pika==1.3.2
python-dotenv==1.0.0
loguru==0.7.2
aiohttp==3.9.1
